+++ packages/playground/optimize-deps/vite.config.js
@@ -9,12 +9,7 @@ module.exports = {
   },
 
   optimizeDeps: {
-    include: [
-      'dep-linked-include',
-      // required since it isn't in node_modules and is ignored by the optimizer otherwise
-      'dep-esbuild-plugin-transform',
-      'nested-exclude>nested-include'
-    ],
+    include: ['dep-linked-include', 'nested-exclude > nested-include'],
     exclude: ['nested-exclude'],
     esbuildOptions: {
       plugins: [