+++ packages/vite/src/node/logger.ts
@@ -1,10 +1,13 @@
 /* eslint no-console: 0 */
 
 import chalk from 'chalk'
-import readline from 'readline'
+import { AddressInfo, Server } from 'net'
 import os from 'os'
+import readline from 'readline'
 import { RollupError } from 'rollup'
-import { Hostname } from './utils'
+import { ResolvedConfig } from '.'
+import { CommonServerOptions } from './http'
+import { Hostname, resolveHostname } from './utils'
 
 export type LogType = 'error' | 'warn' | 'info'
 export type LogLevel = LogType | 'silent'
@@ -63,10 +66,9 @@ export function createLogger(
   const loggedErrors = new WeakSet<Error | RollupError>()
   const { prefix = '[vite]', allowClearScreen = true } = options
   const thresh = LogLevels[level]
-  const clear =
+  const canClearScreen =
     allowClearScreen && process.stdout.isTTY && !process.env.CI
-      ? clearScreen
-      : () => {}
+  const clear = canClearScreen ? clearScreen : () => {}
 
   function output(type: LogType, msg: string, options: LogErrorOptions = {}) {
     if (thresh >= LogLevels[type]) {
@@ -87,17 +89,21 @@ export function createLogger(
       if (options.error) {
         loggedErrors.add(options.error)
       }
-      if (type === lastType && msg === lastMsg) {
-        sameCount++
-        clear()
-        console[method](format(), chalk.yellow(`(x${sameCount + 1})`))
-      } else {
-        sameCount = 0
-        lastMsg = msg
-        lastType = type
-        if (options.clear) {
+      if (canClearScreen) {
+        if (type === lastType && msg === lastMsg) {
+          sameCount++
           clear()
+          console[method](format(), chalk.yellow(`(x${sameCount + 1})`))
+        } else {
+          sameCount = 0
+          lastMsg = msg
+          lastType = type
+          if (options.clear) {
+            clear()
+          }
+          console[method](format())
         }
+      } else {
         console[method](format())
       }
     }
@@ -137,7 +143,37 @@ export function createLogger(
   return logger
 }
 
-export function printServerUrls(
+/**
+ * @deprecated Use `server.printUrls()` instead
+ */
+export function printHttpServerUrls(
+  server: Server,
+  config: ResolvedConfig
+): void {
+  printCommonServerUrls(server, config.server, config)
+}
+
+export function printCommonServerUrls(
+  server: Server,
+  options: CommonServerOptions,
+  config: ResolvedConfig
+): void {
+  const address = server.address()
+  const isAddressInfo = (x: any): x is AddressInfo => x?.address
+  if (isAddressInfo(address)) {
+    const hostname = resolveHostname(options.host)
+    const protocol = options.https ? 'https' : 'http'
+    printServerUrls(
+      hostname,
+      protocol,
+      address.port,
+      config.base,
+      config.logger.info
+    )
+  }
+}
+
+function printServerUrls(
   hostname: Hostname,
   protocol: string,
   port: number,
@@ -153,7 +189,7 @@ export function printServerUrls(
   } else {
     Object.values(os.networkInterfaces())
       .flatMap((nInterface) => nInterface ?? [])
-      .filter((detail) => detail.family === 'IPv4')
+      .filter((detail) => detail && detail.address && detail.family === 'IPv4')
       .map((detail) => {
         const type = detail.address.includes('127.0.0.1')
           ? 'Local:   '